+++ config/const.go
@@ -6,4 +6,9 @@ const (
 	envReliablyOrg              = "RELIABLY_ORG"
 	envReliablyEntityServerHost = "RELIABLY_ENTITY_HOST"
 	envDebug                    = "DEBUG"
+
+	RELIABLY_HOST  = envReliablyHost
+	RELIABLY_ORG   = envReliablyOrg
+	RELIABLY_TOKEN = envReliablyToken
+	DEBUG          = envDebug
 )